Silenced all warnings

+++ pfs/pfs_fio.c
@@ -685,7 +685,7 @@ s64 _seek(pfs_file_slot_t *fileSlot, s64 offset, int whence, int mode)
 		return rv;
 	}
 
-	dprintf("ps2fs: Seek successfull! newPos = %ld\n", (u32)newPos);
+	dprintf("ps2fs: Seek successfull! newPos = %lld\n", newPos);
 
 	return newPos;
 }
